mention profile page on first hit in google on access tokens
This page is the first hit on google when it comes to long lived access token and bearer token for home assistant and it mentions nothing about the profile page being a place to get it when looking for it. 

I know this page is about programmatic access but thought it would be worth mentioning the profile page.